Merged PR 2521: Use the new enqueue route and always load the URL in the WebView
The SDK now uses the following route for API calls: api/mobileapp/queue/{customerId}/{waitingRoomIdOrAlias}/enqueue
Responses are now handled differently.

If a QueueUrl is returned the WebView will display the page.
If the WR has captcha enabled the page would still be shown and the QueueID will be stored in the WebView's cookies.

Note that if outdated SDKs are used, this will lead to the cache to be always empty. The user can still get his position though, since cookies are used. Only the last page won't be remembered.

Related work items: #7489